chore: #75 認証ワークフローの型設計を見直す
