+++ backend/src/domains/user/index.ts
@@ -1,3 +1,5 @@
+import { z } from 'zod'
+
 export type UserData = {
   name: string
   auth0UserId: string
@@ -5,3 +7,14 @@ export type UserData = {
 }
 
 export type User = UserData & { id: string }
+
+export const Auth0UserInfoSchema = z.object({
+  sub: z.string(),
+  nickname: z.string(),
+  name: z.string(),
+  email: z.string(),
+  picture: z.string(),
+  updated_at: z.string(),
+  email_verified: z.boolean(),
+})
+export type Auth0UserInfo = z.infer<typeof Auth0UserInfoSchema>
